+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - Dark mode
 - Changed
   - Display format of tab count and tab count warning for clarity
+  - Improved schema detection
+  - Added schemas to no-lazy-load rule
 
 ## 1.6.0